+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/pocket/PocketPeripheralBuildFunction.kt
@@ -1,4 +1,4 @@
-package site.siredvin.peripheralium.api
+package site.siredvin.peripheralium.api.pocket
 
 import dan200.computercraft.api.pocket.IPocketAccess
 import site.siredvin.peripheralium.api.peripheral.IOwnedPeripheral