+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/pocket/PocketUpgradeContainer.kt
@@ -1,0 +1,8 @@
+package site.siredvin.peripheralium.api.pocket
+
+import dan200.computercraft.api.pocket.IPocketUpgrade
+import dan200.computercraft.api.upgrades.UpgradeData
+
+interface PocketUpgradeContainer {
+    fun getStoredUpgrades(): List<UpgradeData<IPocketUpgrade>>
+}