+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/turtle/TurtlePeripheralBuildFunction.kt
@@ -1,4 +1,4 @@
-package site.siredvin.peripheralium.api
+package site.siredvin.peripheralium.api.turtle
 
 import dan200.computercraft.api.turtle.ITurtleAccess
 import dan200.computercraft.api.turtle.TurtleSide