+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/turtle/TurtleUpgradeContainer.kt
@@ -1,0 +1,8 @@
+package site.siredvin.peripheralium.api.turtle
+
+import dan200.computercraft.api.turtle.ITurtleUpgrade
+import dan200.computercraft.api.upgrades.UpgradeData
+
+interface TurtleUpgradeContainer {
+    fun getStoredUpgrades(): List<UpgradeData<ITurtleUpgrade>>
+}